Use Bridge resource type instead of Transports in rcol
We want to pave the way for importing vanilla bridges into rdsys. This
change will store resources as bridges with a slice of Transports. This
will make it easier for us to test both vanilla bridges and bridges with
transports, and also distribute transports without distributing a more
easily blocked flavour.