Configure what distributor does distribute each resource type
Some distributors distribute only one type of resource, for example
moat does only distribute obfs4 bridges and not vanilla ones. We don't
want to assign resources to a distributor that will not make use of
them.

Let's add an item in the resource definition to configure the
distributors that each resource will support.

Closes: #63