Report bridge status for non vanilla bridges
Transport bridges include the IP address in the hash index, together
with the fingerprint and the type. We can not address them directly just
knowing the fingerprint.

For now let's do a sequential search over the fingerprint, in the future
this might become a problem and we might need to make it more efficient.

Closes: #88